+++ CHANGES.md
@@ -2,8 +2,9 @@
 
 ## Version 4.1.0 (to be released)
 
-- vulnerability scanning using GitHub CodeQL
 - internal logging was migrated from SLF4J to Log4j-api
+- vulnerability scanning using GitHub CodeQL
+- add Cell.setError(), Cell.isError()
 - small fixes and enhancements
 
 ## Version 4.0.3